requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
imbalanced-learn>=0.11
shap>=0.44
matplotlib>=3.7
seaborn>=0.13
pyyaml>=6.0
joblib>=1.3
jupyter>=1.0
notebook>=7.0
ipykernel>=6.0
